+++ tools/measure_RNG_performance.h
@@ -5,8 +5,8 @@ template<typename RNG> double _measure_RNG_performance_16(RNG *rng) {
 	enum {NUM_CLOCKS_TO_TEST = int(CLOCKS_PER_SEC * .5) + 1};
 	//RAW_RNG rng(PractRand::SEED_AUTO);
 	Uint16 buffy[1024];
-	int clock0 = clock();
-	int clock1, clock2;
+	long clock0 = clock();
+	long clock1, clock2;
 	while ((clock1 = clock()) == clock0) ;
 	int j = 0;
 	do {
@@ -29,8 +29,8 @@ template<typename RNG> double _measure_RNG_performance_32(RNG *rng) {
 	enum {NUM_CLOCKS_TO_TEST = int(CLOCKS_PER_SEC * 0.5) + 1};
 	//RAW_RNG rng(PractRand::SEED_AUTO);
 	Uint32 buffy[1024] = {0};
-	int clock0 = clock();
-	int clock1, clock2;
+	long clock0 = clock();
+	long clock1, clock2;
 	while ((clock1 = clock()) == clock0) ;
 	int j = 0;
 	do {
@@ -53,8 +53,8 @@ template<typename RNG> double _measure_RNG_performance_64(RNG *rng) {
 	enum {NUM_CLOCKS_TO_TEST = int(CLOCKS_PER_SEC * 0.5) + 1};
 	//RAW_RNG rng(PractRand::SEED_AUTO);
 	Uint64 buffy[1024] = {0};
-	int clock0 = clock();
-	int clock1, clock2;
+	long clock0 = clock();
+	long clock1, clock2;
 	while ((clock1 = clock()) == clock0) ;
 	int j = 0;
 	do {
